Organised code into classes and files for easier management.
